saving test progress:

+++ weapon.py
@@ -7,6 +7,6 @@ class Weapon(Ability):
     between one half to the full attack power of the weapon.
     """
     def attack(self):
-        divided_number = self.max_damage/2
+        divided_number = int(self.max_damage/2)
         random_interger = random.randint(divided_number, self.max_damage)
         return random_interger
